backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlalchemy
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pillow
numpy
opencv-python
torch
torchvision
stable-baselines3
gymnasium
websockets
openai
requests
pandas
matplotlib
seaborn
plotly
psycopg2-binary